requirements.txt
fastapi==0.109.0
uvicorn==0.26.0
python-multipart==0.0.6
python-dotenv==1.0.0
azure-cognitiveservices-speech==1.34.0
openai==1.59.7
numpy==1.26.3
scipy==1.11.4
scikit-learn==1.4.0
pydantic==2.5.3
python-jose==3.3.0
passlib==1.7.4
sqlalchemy==1.4.50
psycopg2-binary==2.9.9
python-jose[cryptography]==3.3.0
requests==2.31.0
ffmpeg-python==0.2.0
